% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devnews.R
\name{build_news_db_from_file}
\alias{build_news_db_from_file}
\title{Build a news database}
\usage{
build_news_db_from_file(path, mode = c("auto", "Rd", "md", "plain"))
}
\arguments{
\item{path}{Path to the NEWS file}

\item{mode}{Parser to use. \code{"auto"} determines the parser based on the file extension of \code{path}.}
}
\value{
news_db object, or NULL if the parser failed to find any valid entries
}
\description{
Build a news database
}
