% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devnews.R
\name{find_news_file}
\alias{find_news_file}
\title{Search a package source tree for a NEWS file}
\usage{
find_news_file(dir)
}
\arguments{
\item{dir}{Directory to package source tree}
}
\value{
Path to the first matching NEWS file, or NULL if none is found
}
\description{
Follows the search order documented in \code{utils::news()}.
}
