% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devnews.R
\name{show_news}
\alias{show_news}
\title{View NEWS entries as HTML}
\usage{
show_news(news_db, ...)
}
\arguments{
\item{news_db}{NEWS entry database, a news_db object}

\item{...}{Parameters passed to tools::HTMLheader()}
}
\value{
Nothing
}
\description{
Launches a browser to view the HTML, or if in rstudio uses the viewer pane.
}
